move: svg파일 pulic폴더로 이동
